requirements.txt

flask
flask-cors
mysql-connector-python
openai
python-dotenv
streamlit
google-generativeai
twilio
psycopg2-binary